Внести правки согласно комментариям ревью:
- Денежные расчеты производить в BigDecimal:
Обновить DeliveryUtil — все константы теперь BigDecimal
Обновить DeliveryServiceImpl.deliveryCost() — все расчеты используют BigDecimal
Обновить интерфейсы: DeliveryService, DeliveryOperation
Обновить контроллеры: DeliveryController
Обновить сервисы: OrderFacadeImpl, OrderServiceImpl для работы с BigDecimal

- Добавить подробное логирование с ID заказа:
Все логи содержат orderId для идентификации записей в приложении.

+++ commerce/delivery/src/main/java/ru/practicum/controller/DeliveryController.java
@@ -10,6 +10,7 @@ import ru.practicum.dto.DeliveryDto;
 import ru.practicum.dto.OrderDto;
 import ru.practicum.service.DeliveryService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Validated
@@ -46,10 +47,12 @@ public class DeliveryController implements DeliveryClient {
 
     @Override
     @PostMapping("/cost")
-    public Double deliveryCost(@RequestBody @Valid OrderDto orderDto) {
+    public BigDecimal deliveryCost(@RequestBody @Valid OrderDto orderDto) {
         return deliveryService.deliveryCost(orderDto);
     }
 }
+
+
